util: add scratchdir implementation
This adds a new scratchdir module which exposes a single type called
`ScratchDir()`. This is similar to the tempfile.TemporaryDirectory()
python standard library functionality, but it provides integration with
parallel cleanup handlers. In particular, it uses lockfiles to claim a
directory and employs retry-loops wherever necessary.

The idea is to use this as basis for all our internal pipeline
operations. Anything that needs writable trees will have to import the
tree into a ScratchDir. From there, we will then provide a file-system
cache later on, which allows to move data out of ScratchDir objects
into a read-only cache hierarchy.

The nameing is kinda the last idea I could come up with that does not
conflict with any names we used so far:

  * `store` is used by Composer, and it kinda implies long-living data,
    so not very fitting here.
  * `cache` will by used by the actual file-system cache, so cannot use
    it here.
  * `build` / `builddir` conflicts with `build-tree` and
    `build-pipeline`. It has no specific connection to these, so reusing
    the term `build` would be confusing.
  * `tmp` / `tmpdir` is already used by `/tmp`, the standard-library,
    and all official generic directory handlers.

The term `scratch-paper` is already established, so `scratch-dir` seems
suitable for the one-shot operations we do.