util: add file system cache
This commit introduces a new utility module called `fscache`. It
implements a cache module that stores data on the file system. It
supports parallel access and protects data with file-system locks.

This first draft only contains the bare minimum to keep compatibility
to the current ObjectStore. That is, it retains the same file-system
layout and does no automatic cache pruning. However, it does introduce
locking to allow parallel access to the cache.

Furthermore, while this introduces boot-id tagging of cache-entries, it
does not enforce it, yet.

There is lots of room for future improvement. This commit simply
introduces the scaffolding and infrastructure to make parallel cache
accesses possible.