news: reformat according to markdown
The NEWS file currently uses non-markdown indentation, but uses the
markdown file-format. Since both GitHub-Repository-Viewer and the
Release Pages can display proper markdown, lets just re-indent the file
according to the markdown specs.

While at it, also switch out the list-indicator in front of the
timestamp to a proper hyphen, to avoid it being rendered as list.