stages/rpm: drop GPG signature verification
The documentation (wrongly) indicated that only signed RPMs were
verified, but in fact all were, and unsigned RPMs were declined.

We need to support both signed and unsigned RPMs, but rather than adding
support for that, let us reconsider the whole concept.

The manifest specifies the content hash of each RPM, which would include
their signatures. It should be the responsibility of the generator of
the manifest to make sure that the correct RPMs are included. Verifying
the signatures of individual RPMs does not gain us any significant security
properties, as a compromised manifest generation would be susceptible to
roll-back attacks, even if all the RPMs are signed.

The security of image generation is dependent on the security of
manifest generation, so let us not add any additional snake-oil beyond
that.

We still import GPG keys into the rpm database, so they will be
available in the image.

Signed-off-by: Tom Gundersen <teg@jklm.no>